Move NodeID to its own file
