+++ common/types/layer.go
@@ -2,12 +2,9 @@
 package types
 
 import (
-	"encoding/hex"
 	"fmt"
 	"strconv"
 	"sync/atomic"
-
-	"github.com/spacemeshos/ed25519"
 
 	"github.com/spacemeshos/go-spacemesh/log"
 )
@@ -139,43 +136,6 @@ func (l LayerID) Field() log.Field { return log.Uint32("layer_id", l.Value) }
 func (l LayerID) String() string {
 	return strconv.FormatUint(uint64(l.Value), 10)
 }
-
-// BytesToNodeID is a helper to copy buffer into NodeID struct.
-func BytesToNodeID(buf []byte) (id NodeID) {
-	copy(id[:], buf)
-	return id
-}
-
-// ExtractNodeIDFromSig extracts the NodeID from a signature.
-func ExtractNodeIDFromSig(msg, sig []byte) (NodeID, error) {
-	pub, err := ed25519.ExtractPublicKey(msg, sig)
-	if err != nil {
-		return NodeID{}, err
-	}
-	return BytesToNodeID(pub), nil
-}
-
-// NodeID contains a miner's public key.
-type NodeID [32]byte
-
-// String returns a string representation of the NodeID, for logging purposes.
-// It implements the Stringer interface.
-func (id NodeID) String() string {
-	return hex.EncodeToString(id.Bytes())
-}
-
-// Bytes returns the byte representation of the Edwards public key.
-func (id NodeID) Bytes() []byte {
-	return id[:]
-}
-
-// ShortString returns a the first 5 characters of the ID, for logging purposes.
-func (id NodeID) ShortString() string {
-	return Shorten(id.String(), 5)
-}
-
-// Field returns a log field. Implements the LoggableField interface.
-func (id NodeID) Field() log.Field { return log.Stringer("node_id", id) }
 
 // Layer contains a list of proposals and their corresponding LayerID.
 type Layer struct {
